bugs fixed: -m64 knh_printf %d mistyped

+++ konoha-0.5/include/konoha/konoha_class.h
@@ -752,8 +752,8 @@ typedef struct knh_Exception {
 	knh_String_t* message;
 	Object* bag;
 	knh_Array_t* traces;
-	char* file;
-	int line;
+	char*     file;
+	size_t    line;
 } knh_Exception_struct;
 
 /* ------------------------------------------------------------------------ */
